Pass actual addresses to picoquic_prepare_packet
The picoquic_prepare_packet() and picoquic_xia_sendmsg() calls were
being passed storage for the peer and local addresses with the
expectation that these calls will fill in the addresses. However, the
calls don't have the means to recover the addresses because the XIA
addresses are loosely tied to sockfd parameter only in the context of
the caller. Since, we have the addresses readily available from
picoquic_xia_select() call, we just pass them in to future picoquic
calls that need them.

Also fixed the print_address() function that was expecting only an IPv4
or IPv6 address.