Setup fwd table on router via XIAConfigHelper
The XIAConfigHelper application allows routing table entries to be made
over IP, so using that for the overlay network.
The previous iteration of this code was sending a special registration
message with a magic number 0xC0DA that the router was parsing as a
special message outside of the remaining flow. This change makes the
router's work simpler and avoids the need for the new
XIAAIDRoutingTable. That element will be eliminated in a future change.