New addr buffers passed to picoquic_prepare_packet
picoquic_prepare_packet() takes local and peer addresses. We were
passing the buffers that were filled in by picoquic_xia_select(). This
in not correct because picoquic_prepare_packet() should use the correct
addresses corresponding to the connection for which this packet is
being prepared. This change creates new buffers on stack to ensure we
are not reusing incorrect addresses.