picoquic_xia_serve_cid() creates route for CID
New API call picoquic_xia_serve_cid() creates a forwarding table entry
for the given CID on the router. This entry is an IP overlay nexthop.
For now, there is no check to ensure that the given CID is valid.
Eventually, this API will ensure that the CID being advertised is
actually present locally.