Xcache now sending data to client.
With this change, Xcache is able to send data to a client but only
after client makes contact and sends some data. The earlier version
tried to send data simply in response to a client connection but that
results in PICOQUIC_ERROR_INVALID_STREAM_ID from
picoquic_find_stream_for_writing() in response to
picoquic_add_to_stream(). This happens in both ready and almost_ready
state, but doesn't happen after stream_data is received. The
documentation in sample code suggested that it should be possible to
queue up data to send even in almost_ready state but that was not the
case in practice.