When comparing addresses, just compare intents.
Since the DAG may be modified in path, we just need to make sure that
the intent XID in both DAGs matches.

So if a client tries to connect to:
  * -------------------------> CID1
  '------> ADa ---> HIDa ---'

and a response comes back from:
  *--> ADb --> HIDb ---> CID1

It will be an acceptable response. The xcache is responsible for
ensuring that the entire chunk is valid after it has been retrieved.