find_library doesn't work if XIA is not downloaded
When starting a fresh build with cmake, XIA code would not have been
downloaded yet. It is downloaded when make is called to build picoquic.
This causes the find_library() commands for libxcachehdr.so and
libxcache.so to fail on a fresh clone of this repository. This change
should fix that.