Bringing in the xcache interest CID engine
The interest CID handler from old xcache included a thread pool. We
hope to use the same thread pool for all of xcache tasks. So this
change brings in that thread pool along with a number of cleanup
activities in preparation of merging the ICID functionality with the
QUIC based transfer of chunks.