Use Xcache AID DAG for outgoing ICID requests
This change required the XcacheQUICServer object to own the
QUICXIASocket. Makes Xcache main loop even more simpler. Just create an
AID for the Xcache and give it to the constructor for XcacheQUICServer.
The server will automaticaly create a QUICXIASocket with the AID as the
intent. The corresponding sockfd is then exposed to the Xcache main loop
so it can select() on it.